Fixed 945800 - remove unmaintained code from build step
Bixie components are not deployed, tested, or maintained. They're
also blocked on a few things outside our control. If they become
unblocked it's trivial to create an inverse of this patch and restore
them to the repository. Until then, we can save some complexity and
time off every build by removing them.