fixes bug 965558 by deleting a bunch of code
playdoh left a bunch of files in webapp-django that we don't use
or need. some of them conflict with other infra provided by the parent
repo, so it's best to remove them and avoid confusion.

exempting `webapp-django/lib` because funfactory blows up if that magic
file is missing. when/if we move off playdoh, though, I'll be waiting
to reclaim those bits.